enhance(bulk fastt): add is_empty, is_not_empty
