+++ adminSiteClient/VariablesAnnotationTypesAndUtils.tsx
@@ -2,21 +2,50 @@ import { queryParamsToStr } from "../clientUtils/urls/UrlUtils.js"
 import {
     BinaryLogicOperation,
     BinaryLogicOperators,
-    BooleanOperation,
+    BooleanAtom,
+    ComparisonOperator,
+    EqualityComparision,
+    EqualityOperator,
+    JsonPointerSymbol,
+    Negation,
+    NumberAtom,
+    NumericComparision,
     Operation,
+    SqlColumnName,
     StringAtom,
     StringContainsOperation,
+    NullCheckOperation,
+    NullCheckOperator,
     StringOperation,
+    BooleanOperation,
 } from "../clientUtils/SqlFilterSExpression.js"
-import React from "react"
-import { IconDefinition } from "@fortawesome/fontawesome-common-types/index.js"
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome/index.js"
+import * as React from "react"
+import { IconDefinition } from "@fortawesome/fontawesome-common-types"
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 
 import { GrapherInterface } from "../grapher/core/GrapherInterface.js"
 import {
     VariableAnnotationsResponseRow,
     VariableAnnotationPatch,
 } from "../clientUtils/AdminSessionTypes.js"
+import AntdConfig from "react-awesome-query-builder/lib/config/antd"
+import { BasicConfig, SimpleField } from "react-awesome-query-builder"
+import { Utils as QbUtils } from "react-awesome-query-builder"
+// types
+import {
+    JsonGroup,
+    JsonItem,
+    Config,
+    ImmutableTree,
+} from "react-awesome-query-builder"
+import { match } from "ts-pattern"
+import { excludeUndefined, isArray } from "../clientUtils/Util.js"
+import { isPlainObject } from "lodash"
+import {
+    EditorOption,
+    FieldDescription,
+} from "../clientUtils/schemaProcessing.js"
+
 export function parseVariableAnnotationsRow(
     row: VariableAnnotationsResponseRow
 ): VariableAnnotationsRow {
@@ -245,4 +274,209 @@ export const getItemStyle = (
 
 export function isConfigColumn(columnName: string): boolean {
     return columnName.startsWith("/")
+}
+
+export const filterPanelInitialConfig: BasicConfig = AntdConfig as BasicConfig
+export const initialFilterQueryValue: JsonGroup = {
+    id: QbUtils.uuid(),
+    type: "group",
+}
+export type FilterPanelState = {
+    tree: ImmutableTree
+    config: Config
+}
+
+export function getLogicOperator(str: string): BinaryLogicOperators {
+    if (str === "AND") return BinaryLogicOperators.and
+    else if (str === "OR") return BinaryLogicOperators.or
+    else throw Error(`unknown logic operator: ${str}`)
+}
+
+export function getComparisonOperator(
+    str: string
+): ComparisonOperator | undefined {
+    return match(str)
+        .with("less", () => ComparisonOperator.less)
+        .with("less_or_equal", () => ComparisonOperator.lessOrEqual)
+        .with("greater", () => ComparisonOperator.greater)
+        .with("greater_or_equal", () => ComparisonOperator.greaterOrEqual)
+        .otherwise(() => undefined)
+}
+
+export function getNullCheckOperator(
+    str: string
+): NullCheckOperator | undefined {
+    return match(str)
+        .with("is_null", () => NullCheckOperator.isNull)
+        .with("is_not_null", () => NullCheckOperator.isNotNull)
+        .otherwise(() => undefined)
+}
+
+export function getFieldSymbol(fieldName: string): Operation {
+    if (isConfigColumn(fieldName)) return new JsonPointerSymbol(fieldName)
+    else
+        return new SqlColumnName(
+            readOnlyColumnNamesFields.get(fieldName)!.sExpressionColumnTarget
+        )
+}
+
+export function getValueAtom(val: any): Operation | undefined {
+    if (typeof val === "string") return new StringAtom(val)
+    else if (typeof val === "number") return new NumberAtom(val)
+    else if (typeof val === "boolean") return new BooleanAtom(val)
+    else return undefined
+}
+
+export function getEqualityOperator(str: string): EqualityOperator | undefined {
+    if (str === "equal" || str === "select_equals")
+        return EqualityOperator.equal
+    else if (str === "not_equal" || str === "select_not_equals")
+        return EqualityOperator.unequal
+    else return undefined
+}
+
+export function filterTreeToSExpression(
+    filterTree: JsonItem
+): Operation | undefined {
+    if (filterTree.type === "group") {
+        const logicOperator = getLogicOperator(
+            filterTree.properties?.conjunction ?? "AND"
+        )
+        let children: Operation[] = []
+        if (isArray(filterTree.children1))
+            children = excludeUndefined(
+                filterTree.children1?.map(filterTreeToSExpression)
+            )
+        else if (isPlainObject(filterTree.children1))
+            children = excludeUndefined(
+                Object.values(
+                    filterTree.children1 as Record<string, JsonItem>
+                ).map(filterTreeToSExpression)
+            )
+        else if (filterTree.children1 !== undefined)
+            console.warn("unexpected content of children1")
+        if (filterTree.children1 === undefined || children.length === 0)
+            return undefined
+        const operation = new BinaryLogicOperation(logicOperator, children)
+        if (filterTree.properties?.not) return new Negation(operation)
+        else return operation
+    } else if (filterTree.type === "rule") {
+        return match(filterTree.properties.operator)
+            .when(
+                (op) => op && getComparisonOperator(op),
+                (op) => {
+                    const operator = getComparisonOperator(op as string)
+                    const field = getFieldSymbol(filterTree.properties.field!)
+                    if (
+                        filterTree.properties.value.length === 0 ||
+                        filterTree.properties.value[0] === undefined
+                    )
+                        return undefined
+                    const val = getValueAtom(filterTree.properties.value[0])
+                    return new NumericComparision(operator!, [field, val!])
+                }
+            )
+            .when(
+                (op) => op && getEqualityOperator(op),
+                (op) => {
+                    const operator = getEqualityOperator(op as string)
+                    const field = getFieldSymbol(filterTree.properties.field!)
+                    const val = getValueAtom(filterTree.properties.value[0])
+                    if (val === undefined) return undefined
+                    return new EqualityComparision(operator!, [field, val])
+                }
+            )
+            .when(
+                (op) => op && op === "like",
+                () => {
+                    const field = getFieldSymbol(filterTree.properties.field!)
+                    if (
+                        filterTree.properties.value.length === 0 ||
+                        filterTree.properties.value[0] === undefined
+                    )
+                        return undefined
+                    const val = new StringAtom(filterTree.properties.value[0])
+                    return new StringContainsOperation(field, val)
+                }
+            )
+            .when(
+                (op) => op && getNullCheckOperator(op as string),
+                (op) => {
+                    const operator = getNullCheckOperator(op as string)!
+                    const field = getFieldSymbol(filterTree.properties.field!)
+                    return new NullCheckOperation(operator!, field)
+                }
+            )
+            .with("is_empty", "is_not_empty", (operator) => {
+                const field = getFieldSymbol(filterTree.properties.field!)
+                const op: EqualityOperator = match(operator)
+                    .with("is_empty", () => EqualityOperator.equal)
+                    .with("is_not_empty", () => EqualityOperator.unequal)
+                    .exhaustive()
+                return new EqualityComparision(op, [field, new StringAtom("")])
+            })
+
+            .otherwise(() => undefined)
+    }
+    return undefined
+}
+
+export function simpleColumnToFilterPanelFieldConfig(
+    column: ReadOnlyColumn
+): [string, SimpleField] {
+    const fieldType = match(column.type)
+        .with("string", () => "text")
+        .with("number", () => "number")
+        .with("datetime", () => "datetime")
+        .exhaustive()
+
+    return [
+        column.key,
+        {
+            label: column.key,
+            type: fieldType,
+            valueSources: ["value"],
+            //preferWidgets: widget [widget],
+        },
+    ]
+}
+
+export function fieldDescriptionToFilterPanelFieldConfig(
+    description: FieldDescription
+): [string, SimpleField] | undefined {
+    const widget = match(description.editor)
+        .with(EditorOption.checkbox, () => "boolean")
+        .with(EditorOption.colorEditor, () => undefined)
+        .with(EditorOption.dropdown, () => "select")
+        .with(EditorOption.mappingEditor, () => undefined)
+        .with(EditorOption.numeric, () => "number")
+        .with(EditorOption.primitiveListEditor, () => undefined)
+        .with(EditorOption.textarea, () => "text")
+        .with(EditorOption.textfield, () => "text")
+        .exhaustive()
+
+    if (widget !== undefined)
+        return [
+            description.pointer,
+            {
+                label: description.pointer,
+                type: widget,
+                valueSources: ["value"],
+                //preferWidgets: widget [widget],
+                fieldSettings: {
+                    listValues: description.enumOptions,
+                },
+            },
+        ]
+    else return undefined
+}
+
+export function getFinalConfigLayerForVariable(id: number) {
+    return {
+        version: 1,
+        dimensions: [{ property: "y", variableId: id }],
+        map: {
+            variableId: id,
+        },
+    }
 }
